Move the local time display to a dedicated banner
